building_tools (dist): Add NPE check in EDT along with a defensive copy
See #21833: NPE: Cannot invoke "org.openstreetmap.josm.data.coor.EastNorth.east()".
This was caused by something either calling the reset method *or* one of the
methods for calculating the EastNorth returning null.

In any case, we fixed the potential race condition by making a defensive copy
of the en array, and then checking the defensive copy for nulls in the EDT,
where appropriate.


git-svn-id: https://josm.openstreetmap.de/osmsvn@35909 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4