http2 (dist): see #21906: Properly use user set User-Agent header
This also enables the http2 plugin to be disabled without JOSM restart -- we
just set the HttpClient factory back to Http1Client::new. Ideally, users would
still restart JOSM to ensure that all classes are unloaded, but this should
make it easier to debug if an issue is due to http2 or something else.



git-svn-id: https://josm.openstreetmap.de/osmsvn@35922 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4