Fix #17465: IAE: Listener buildings_tools.DrawBuildingAction was not registered before or already removed.
This occurs when a user attempts to enter the mapmode when the edit dataset is
null. We fix the problem by (a) updating the `layerIsSupported` method to use
`isEditableDataLayer` (introduced in r13434) and (b) removing a check that used
to ensure that if a user entered the mapmode when there was no editable layer,
we immediately exit from the mapmode and enter into the select mode. As all
calls to `enterMode` should be guarded by `layerIsSupported` (on layer change),
this is safe now that we are checking if the layer is editable.


git-svn-id: https://josm.openstreetmap.de/osmsvn@35997 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4