Fix #22520: IllegalStateException: Missing merge target for node (patch by GerdP, modified to add tests)
The added test is essentially an integration test instead of a unit test, mostly
due to having no good way to do a test specifically for the method at issue.
The test uses wiremock to (a) speed up the test and (b) avoid adding a test-time
dependency on the OSM API servers. The test required a custom wiremock response
transformer for the OSM multi-fetch API, since it is not guaranteed that the
order of objects will remain the same, which will lead to stub misses, failing
the test.

git-svn-id: https://josm.openstreetmap.de/osmsvn@36042 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4