See #23227: Add layer option for Bing Imagery
iD/Rapid use the layer `AerialOSM` instead of `Aerial`, and that layer is higher
resolution (but older) in some locations. This adds a method to set the Bing Maps
layer (see https://learn.microsoft.com/en-us/bingmaps/rest-services/imagery/get-imagery-metadata#template-parameters )
so that downstream users can decide whether their agreement with Bing Maps allows
the use of `AerialOSM`. We have not received communication as to which layer Bing
wants us to use for editing OpenStreetMap.

git-svn-id: https://josm.openstreetmap.de/osmsvn@36174 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4